+++ app/extensions.py
@@ -7,6 +7,7 @@ Each extension is initialized in the app factory located in app.py.
 from flask import Flask
 from flask_caching import Cache
 from flask_debugtoolbar import DebugToolbarExtension
+from flask_login import LoginManager
 from flask_mail import Mail
 from flask_migrate import Migrate
 from flask_sqlalchemy import SQLAlchemy
@@ -18,6 +19,7 @@ cache = Cache()
 db = SQLAlchemy()
 migrate = Migrate()
 mail = Mail()
+login_manager = LoginManager()
 
 
 def init_app(app: Flask) -> None:
@@ -31,3 +33,4 @@ def init_app(app: Flask) -> None:
     db.init_app(app)
     migrate.init_app(app, db)
     mail.init_app(app)
+    login_manager.init_app(app)