+++ CHANGELOG.md
@@ -15,4 +15,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Adds mandatory email field for registration
 
 ### Changed
+- Adds `Flask-Login` for handling logins
 - Moves database over to `SQLAlchemy`